This change will allow us to use parallel execution of test cases. Its working this way:
Changes made to substeps maven plugin to analyze all feature files in relevant directory and extract the first execution tag from them (example: @LDY-123). 

My logic is that this tag should be unique and used in separate threads for executing the test which have it. 
If the tag is not unique all tests that have this tag will be executed one after each other.
Reporting for parallel test for now is done in separate folder for each tread (each tread will create its own report).

For substeps execution in parallel we need to have this tags in pom.xml for the project which contain the tests: 

		<outputDirectory_tmp>${project.build.directory}</outputDirectory_tmp>
		<threadPoolSize>${pool.size}</threadPoolSize>
		<maxExecutionTime>${max.execution.time.hours}</maxExecutionTime>
		<paralelExecution>${paralel.execution}</paralelExecution> 

First parameter is the directory where report for each thread will be created.
Second one is the size of the tread pool which will be used to execute the test. It actually mean how much treads will be run simultaneously.
Third one is the time we have for execution and if its exceeded all threads should be closed. 
Last one is a flag which if true will use the parallel execution and if not will execute the tests the old way.